docs: re-introduce APIs section
It seems like the APIs section has been removed in 3de3e8c95c73.
However I found it quite handy to be able to quickly access a
webpage describing the API.

It appears like there were two issues with the previous approach:
the OpenAPI schema was locally duplicated (making it often times
outdated and cumbersome to keep in sync), and the swagger-ui
integration needed a dedicated Hugo plugin.

Directly download the OpenAPI schema from GitHub instead of keeping
a copy, and use a HTML template for swagger-ui instead of a plugin.
This should make it easier to maintain.